\name{gose_func}
\alias{gose_func}
\title{Get GOSE data}
\usage{
gose_func(data)
}

\description{
Get the data related to GOSE by just supplying the raw redcap data. You can get the raw redcap data
by using pull_redcap_data function.
}

\arguments{
  \item{data}{
  Provide the raw redcap data obtained from pull_redcap_data function
  }

}
