\name{mrs_func}
\alias{mrs_func}
\title{Get MRS data}
\usage{
mrs_func(data)
}

\description{
Get the data related to MRS by just supplying the raw redcap data. You can get the raw redcap data
by using pull_redcap_data function.
}

\arguments{
  \item{data}{
  Provide the raw redcap data obtained from pull_redcap_data function
  }

}
